Add global MIDI log store and show count
