+++ src/FloatingActionBar.tsx
@@ -1,9 +1,14 @@
 import { useState, useEffect } from 'react';
 import MidiLogger from './MidiLogger';
+import { useMidi } from './useMidi';
+import { useLogStore } from './logStore';
 
 export default function FloatingActionBar() {
   const [showLogger, setShowLogger] = useState(false);
   const [isScrolled, setIsScrolled] = useState(false);
+  const logCount = useLogStore((s) => s.logs.length);
+  const addMessage = useLogStore((s) => s.addMessage);
+  const { listen } = useMidi();
 
   useEffect(() => {
     const handleScroll = () => {
@@ -14,15 +19,20 @@ export default function FloatingActionBar() {
     return () => window.removeEventListener('scroll', handleScroll);
   }, []);
 
+  useEffect(() => {
+    const unlisten = listen(addMessage);
+    return unlisten;
+  }, [listen, addMessage]);
+
   return (
     <>
       <div className={`floating-action-bar ${isScrolled ? 'scrolled' : ''}`}>
-        <button 
+        <button
           className="retro-button"
           onClick={() => setShowLogger(!showLogger)}
           title="Toggle MIDI Logger"
         >
-          {showLogger ? 'HIDE LOG' : 'MIDI LOG'}
+          {showLogger ? 'HIDE LOG' : 'MIDI LOG'} ({logCount})
         </button>
       </div>
       {showLogger && <MidiLogger onClose={() => setShowLogger(false)} />}
